// swiyu-ecosystem-portal-ui/src/app/api/generated/model/protected-verification-submission.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Contact} from './contact';

export interface ProtectedVerificationSubmission {
  id: string;
  partnerId: string;
  sbnId: string;
  entityName: string;
  uid?: string;
  contactPerson?: Contact;
  category: ProtectedVerificationSubmission.CategoryEnum;
  reason: string;
  status: ProtectedVerificationSubmission.StatusEnum;
  submittedAt?: string;
}
export namespace ProtectedVerificationSubmission {
  export const CategoryEnum = {
    PersonalAdministrativeNumber: 'PERSONAL_ADMINISTRATIVE_NUMBER'
  } as const;
  export type CategoryEnum = (typeof CategoryEnum)[keyof typeof CategoryEnum];
  export const StatusEnum = {
    Submitted: 'SUBMITTED',
    Approved: 'APPROVED',
    Rejected: 'REJECTED'
  } as const;
  export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum];
}